+++ etils/ecolab/__init__.py
@@ -25,11 +25,11 @@ from etils.ecolab.highlight_util import highlight_html
 from etils.ecolab.inplace_reload import ReloadMode
 from etils.ecolab.inspects.auto_utils import auto_inspect
 from etils.ecolab.inspects.core import inspect
-from etils.ecolab.module_utils import clear_cached_modules
 from etils.ecolab.patch_utils import patch_graphviz
 from etils.ecolab.patch_utils import set_verbose
 from etils.ecolab.pyjs_com import js_import as pyjs_import
 from etils.ecolab.pyjs_com import register_js_fn
+from etils.epy.adhoc_utils.module_utils import clear_cached_modules
 
 # Activate auto-display by default
 auto_display()